fix(slack): scope context by workspace, time out API calls, share event schema
Filter resolveSlackContext on the integration externalId so a webhook that
already knows the Slack team id cannot pick up another workspace's bot token.
Add an AbortSignal timeout to every SlackClient call so a stalled Slack request
cannot hang the handler. Move the Slack event schema into @orbit/shared so the
webhook route parses request input with a shared validator.
